modified:   app/add_booking_view.php 	modified:   app/add_booking_view.php~ 	modified:   app/add_user_view.php 	modified:   app/add_user_view.php~ 	new file:   app/bookings_moi.php 	new file:   app/bookings_moi.php~ 	modified:   app/cars_list.php 	modified:   app/cars_list.php~ 	modified:   app/index.php 	modified:   app/index.php~ 	new file:   app/trip_view.php 	new file:   app/trip_view.php~ 	modified:   css/style.css 	modified:   css/style.css~ 	modified:   scripts/add_booking.php 	modified:   scripts/add_user.script.php

+++ app/cars_list.php
@@ -25,7 +25,12 @@ $db = mysql_select_db(DB_NAME, $connect) or die('cant select database: '.mysql_e
 		}else{
 			echo "<td>".$s['ownership']."</td>";
 		}
+		
+		if ($_SESSION['privileges'] == "admin"){
 		echo "<td><a href='../scripts/delete_cars.php?cars_id=".$s['cars_id']."'>delete</a></td>";
+		}else{
+		echo '<td>&nbsp;</td>';
+		}
 		echo "</tr>";  
 		echo "<tr id='blank' ><td colspan='4'>&nbsp</td></tr>";           	
 	}
